refactor(occt): extract face/edge meshing into shared meshing.ts
Pre-condition for v0.21.1 Node-side per-feature meshing pipeline.
Worker behavior unchanged — exercised by existing worker tests.
